requirements.txt
configparser
pillow
tqdm
ordered-set
nuitka
